# openapi.yaml
openapi: 3.0.3
info:
  title: Thai Accounting ERP API
  description: |
    RESTful API for Thai Accounting ERP System (โปรแกรมบัญชีมาตรฐานไทย)
    
    ## Features
    - Double-entry bookkeeping
    - Thai VAT (7%) and withholding tax support
    - Multi-currency support
    - Inventory management with WAC costing
    - Fixed assets with depreciation
    - Payroll with social security calculations
    
    ## Authentication
    The API uses session-based authentication via NextAuth.js.
    Include the session cookie in your requests.
    
    ## API Versions
    - **v1** (stable): `/api/v1/*` - Base functionality
    - **v2** (stable): `/api/v2/*` - Enhanced with GraphQL, field selection, cursor pagination
    - **Latest**: Use `/api/*` for the most recent stable version
    
    ## Rate Limiting
    - Authentication endpoints: 5 requests per minute
    - Other endpoints: 100 requests per minute
    - Headers: `X-RateLimit-Limit`, `X-RateLimit-Remaining`, `X-RateLimit-Reset`
  version: 1.0.0
  contact:
    name: API Support
    email: support@thaiaccounting.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Local development
  - url: http://localhost:3000/api/v1
    description: Version 1 (Stable)
  - url: http://localhost:3000/api/v2
    description: Version 2 (Latest with GraphQL)
  - url: http://localhost:3000/api/graphql
    description: GraphQL Endpoint
...
paths:
...
  /admin/webhooks:
    get:
      tags: [Webhooks]
      summary: List webhooks
      description: Get all webhook subscriptions (Admin only)
      responses:
        '200':
          description: List of webhooks
    post:
      tags: [Webhooks]
      summary: Create webhook
      description: Create a new webhook subscription (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWebhookRequest'
      responses:
        '201':
          description: Webhook created
...
components:
  schemas:
...
    CreateWebhookRequest:
      type: object
      required: [name, url, events]
      properties:
        name:
          type: string
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
            enum:
              - INVOICE_CREATED
              - INVOICE_UPDATED
              - INVOICE_PAID
              - INVOICE_VOIDED
              - RECEIPT_CREATED
              - PAYMENT_CREATED
              - JOURNAL_ENTRY_POSTED
        secret:
          type: string
...
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth session cookie
security:
  - sessionAuth: []